New structure of project [2/3]
todo:
- move src/bot/messages/messager.rs to src/bot/messages.rs (with own function handler to easily add any message-related handler here)
