+++ src/bot/callbacks/mod.rs
@@ -1,5 +1,54 @@
+use crate::bot::callbacks::cobalt_pagination::handle_cobalt_pagination;
+use crate::bot::callbacks::delete::delete_message_handler;
+use crate::bot::callbacks::module::{
+    module_option_handler, module_select_handler, module_toggle_handler, settings_back_handler,
+    settings_set_handler,
+};
+use crate::bot::callbacks::translate::handle_translate_callback;
+use crate::bot::callbacks::whisper::handle_whisper_callback;
+use crate::config::Config;
+use crate::errors::MyError;
+use crate::util::transcription::{back_handler, summarization_handler};
+use std::sync::Arc;
+use teloxide::Bot;
+use teloxide::prelude::{CallbackQuery, Requester};
+
 pub mod cobalt_pagination;
 pub mod delete;
 pub mod module;
 pub mod translate;
 pub mod whisper;
+
+pub async fn callback_query_handlers(bot: Bot, q: CallbackQuery) -> Result<(), MyError> {
+    let config = Arc::new(Config::new().await);
+
+    if let Some(data) = &q.data {
+        if data.starts_with("settings_set:") {
+            settings_set_handler(bot, q).await?
+        } else if data.starts_with("delete_msg") {
+            delete_message_handler(bot, q).await?
+        } else if data.starts_with("summarize") {
+            summarization_handler(bot, q, &config).await?
+        } else if data.starts_with("back_to_full") {
+            back_handler(bot, q, &config).await?
+        } else if data.starts_with("module_select:") {
+            module_select_handler(bot, q).await?
+        } else if data.starts_with("module_toggle") {
+            module_toggle_handler(bot, q).await?
+        } else if data.starts_with("module_opt:") {
+            module_option_handler(bot, q).await?
+        } else if data.starts_with("settings_back:") {
+            settings_back_handler(bot, q).await?
+        } else if data.starts_with("whisper") {
+            handle_whisper_callback(bot, q, &config).await?
+        } else if data.starts_with("tr_") {
+            handle_translate_callback(bot, q, &config).await?
+        } else if data.starts_with("cobalt:") {
+            handle_cobalt_pagination(bot, q, config).await?
+        } else {
+            bot.answer_callback_query(q.id).await?;
+        }
+    }
+
+    Ok(())
+}
