+++ src/bot/dispatcher.rs
@@ -1,7 +1,7 @@
-use crate::bot::handlers::commander::command_handlers;
-use crate::bot::handlers::markups::markuper::callback_query_handlers;
-use crate::bot::handlers::messages::chat::handle_bot_added;
-use crate::bot::handlers::messages::messager::{handle_currency, handle_speech};
+use crate::bot::callbacks::callback_query_handlers;
+use crate::bot::commands::command_handlers;
+use crate::bot::messages::chat::handle_bot_added;
+use crate::bot::messages::messager::{handle_currency, handle_speech};
 use crate::bot::inlines::cobalter::{handle_cobalt_inline, is_query_url};
 use crate::bot::inlines::currency::handle_currency_inline;
 use crate::bot::inlines::whisper::{handle_whisper_inline, is_whisper_query};
