--- src/bot/handlers/commands/mod.rs
@@ -1,4 +1,0 @@
-pub mod settings;
-pub mod speech_recognition;
-pub mod start;
-pub mod translate;
