--- src/bot/handlers/markups/mod.rs
@@ -1,1 +1,0 @@
-pub mod markuper;
