+++ src/bot/handlers/speech_recognition.rs
