+++ src/bot/keyboards/mod.rs
@@ -1,0 +1,2 @@
+pub mod translate;
+pub mod cobalt;
