+++ src/bot/mod.rs
@@ -1,4 +1,7 @@
 mod callbacks;
+pub mod commands;
 pub mod dispatcher;
 mod handlers;
 pub mod inlines;
+pub mod keyboards;
+pub mod messages;
