+++ src/core/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod currency;
+pub mod speech_recognition;
+pub mod translation;
