+++ src/core/services/translation.rs
@@ -1,0 +1,55 @@
+pub const SUPPORTED_LANGUAGES: &[(&str, &str)] = &[
+    ("uk", "🇺🇦 Українська"),
+    ("en", "🇬🇧 English"),
+    ("us", "🇺🇸 English (US)"),
+    ("ru", "🇷🇺 Русский"),
+    ("de", "🇩🇪 Deutsch"),
+    ("fr", "🇫🇷 Français"),
+    ("es", "🇪🇸 Español"),
+    ("it", "🇮🇹 Italiano"),
+    ("zh", "🇨🇳 中文"),
+    ("ja", "🇯🇵 日本語"),
+    ("ko", "🇰🇷 한국어"),
+    ("pl", "🇵🇱 Polski"),
+    ("ar", "🇸🇦 العربية"),
+    ("pt", "🇵🇹 Português"),
+    ("tr", "🇹🇷 Türkçe"),
+    ("nl", "🇳🇱 Nederlands"),
+    ("sv", "🇸🇪 Svenska"),
+    ("no", "🇳🇴 Norsk"),
+    ("da", "🇩🇰 Dansk"),
+    ("fi", "🇫🇮 Suomi"),
+    ("el", "🇬🇷 Ελληνικά"),
+    ("he", "🇮🇱 עברית"),
+    ("hi", "🇮🇳 हिन्दी"),
+    ("id", "🇮🇩 Indonesia"),
+    ("vi", "🇻🇳 Tiếng Việt"),
+    ("th", "🇹🇭 ภาษาไทย"),
+    ("cs", "🇨🇿 Čeština"),
+    ("hu", "🇭🇺 Magyar"),
+    ("ro", "🇷🇴 Română"),
+    ("bg", "🇧🇬 Български"),
+    ("sr", "🇷🇸 Српски"),
+    ("hr", "🇭🇷 Hrvatski"),
+    ("sk", "🇸🇰 Slovenčina"),
+    ("sl", "🇸🇮 Slovenščina"),
+    ("lt", "🇱🇹 Lietuvių"),
+    ("lv", "🇱🇻 Latviešu"),
+    ("et", "🇪🇪 Eesti"),
+];
+pub const LANGUAGES_PER_PAGE: usize = 6;
+
+pub fn normalize_language_code(lang: &str) -> String {
+    match lang.to_lowercase().as_str() {
+        "ua" | "ukrainian" | "украинский" | "uk" => "uk".to_string(),
+        "ru" | "russian" | "русский" => "ru".to_string(),
+        "en" | "english" | "английский" => "en".to_string(),
+        "de" | "german" | "немецкий" => "de".to_string(),
+        "fr" | "french" | "французский" => "fr".to_string(),
+        "es" | "spanish" | "испанский" => "es".to_string(),
+        "it" | "italian" | "итальянский" => "it".to_string(),
+        "zh" | "chinese" | "китайский" => "zh".to_string(),
+        "ja" | "japanese" | "японский" => "ja".to_string(),
+        _ => lang.to_lowercase(),
+    }
+}
